+++ common/src/main/java/io/github/mortuusars/exposure/world/level/storage/ExpectedExposure.java
@@ -5,7 +5,7 @@ import net.minecraft.server.level.ServerPlayer;
 import java.util.function.BiConsumer;
 
 public record ExpectedExposure(String id, long timeoutAt, BiConsumer<ServerPlayer, String> onReceived) {
-    public boolean isTimedOut(long currentUnixTime) {
-        return currentUnixTime > timeoutAt;
+    public boolean isTimedOut(long currentUnixTimeSeconds) {
+        return currentUnixTimeSeconds > timeoutAt;
     }
 }